+++ src/components/auth/LoginForm.tsx
@@ -1,72 +1,21 @@
-import React, { useState } from "react";
-import { useNavigate } from "react-router-dom";
 import BALLFiNLogo from "../../assets/BALLFiN.svg";
 import { Mail, Lock, Eye, EyeOff } from "lucide-react";
-import { login } from "../../api/auth/loginApi";
 import Toast from "@/components/common/Toast";
+import { useLoginForm } from "../../features/auth/useLoginForm";
 
 export default function LoginForm() {
-  const navigate = useNavigate();
-  const [showPassword, setShowPassword] = useState(false);
-  const [formData, setFormData] = useState({
-    email: "",
-    password: "",
-  });
-  const [toast, setToast] = useState<{
-    show: boolean;
-    message: string;
-    type: "success" | "error";
-  }>({ show: false, message: "", type: "success" });
-
-  const handleSubmit = async (e: React.FormEvent) => {
-    e.preventDefault();
-
-    try {
-      const token = await login(formData);
-
-      // 토큰을 로컬 스토리지에 저장
-      localStorage.setItem("access_token", token);
-
-      setToast({
-        show: true,
-        message: "로그인에 성공했습니다.",
-        type: "success",
-      });
-      // 로그인 성공 후 홈페이지로 이동
-      navigate("/");
-      window.location.reload();
-    } catch (error) {
-      console.error("로그인 페이지 에러:", error);
-      let errorMessage = "로그인에 실패했습니다.";
-
-      if (error instanceof Error) {
-        if (error.message === "이메일 또는 비밀번호 오류") {
-          errorMessage = "이메일 또는 비밀번호가 올바르지 않습니다.";
-        } else {
-          errorMessage = error.message;
-        }
-      }
-
-      setToast({
-        show: true,
-        message: errorMessage,
-        type: "error",
-      });
-    }
-  };
-
-  const handleGoogleLogin = () => {
-    console.log("구글 로그인 시도");
-    // 구글 로그인 로직 구현
-  };
-
-  const handleChange = (e: React.ChangeEvent<HTMLInputElement>) => {
-    const { name, value } = e.target;
-    setFormData((prev) => ({
-      ...prev,
-      [name]: value,
-    }));
-  };
+  const {
+    formData,
+    showPassword,
+    toast,
+    handleSubmit,
+    handleGoogleLogin,
+    handleChange,
+    togglePasswordVisibility,
+    closeToast,
+    navigateToSignUp,
+    navigateToForgotPassword,
+  } = useLoginForm();
 
   return (
     <div className="w-full lg:w-1/2 p-8">
@@ -128,7 +77,7 @@ export default function LoginForm() {
             />
             <button
               type="button"
-              onClick={() => setShowPassword(!showPassword)}
+              onClick={togglePasswordVisibility}
               className="absolute inset-y-0 right-0 pr-3 flex items-center"
             >
               {showPassword ? (
@@ -173,14 +122,14 @@ export default function LoginForm() {
           <button
             type="button"
             className="text-[#0A5C2B] hover:underline"
-            onClick={() => navigate("/signup")}
+            onClick={navigateToSignUp}
           >
             회원가입
           </button>
           <button
             type="button"
             className="text-gray-600 hover:underline"
-            onClick={() => navigate("/forgot-password")}
+            onClick={navigateToForgotPassword}
           >
             비밀번호를 잊으셨나요?
           </button>
@@ -202,11 +151,7 @@ export default function LoginForm() {
       </div>
 
       {toast.show && (
-        <Toast
-          message={toast.message}
-          type={toast.type}
-          onClose={() => setToast({ ...toast, show: false })}
-        />
+        <Toast message={toast.message} type={toast.type} onClose={closeToast} />
       )}
     </div>
   );
